# backend/positions-service/openapi.yaml
openapi: 3.0.3
info:
  title: skyPulse positions-service API
  description: >-
    Контракт того, что `positions-service` **реально отдаёт сегодня**. 

  
    Обновление карты — поллинг `GET /api/flights/live` с клиента. Чаще одного
    раза в 5 секунд смысла нет: пайплайн обновляет позиции раз в 15 секунд.
    Плавность движения маркеров добирается экстраполяцией по `velocity`
    и `trueTrack`, а не частотой запросов.


    Единицы: время — unix seconds в UTC, высота — метры, скорость — м/с,
    курс — градусы от севера по часовой стрелке.
  version: 1.0.0

servers:
  - url: /api

tags:
  - name: flights
    description: Борта, позиции, треки
  - name: airports
    description: Справочник аэропортов OurAirports
  - name: pipeline
    description: Состояние поставщика данных

paths:
  /flights/live:
    get:
      tags: [flights]
      summary: Борта для отрисовки маркеров на карте
      description: >-
        Основной запрос карты. Возвращает **плоский массив** позиций, без
        конверта и без метаданных о свежести — за свежестью идти
        в `GET /pipeline-status`.


        Возвращаются **все** борта, попавшие в область, — искусственного
        потолка на размер ответа нет. Без bbox на мировом зуме это порядка
        10–15 тысяч объектов, и прореживать их, если нужно, клиент решает сам.


        Единственная отсечка — по свежести: борт, о котором не было позиции
        дольше `YT_MAX_POSITION_AGE_SECONDS` (по умолчанию 300 с), из выдачи
        пропадает. Она обязательна, потому что `positions_current` хранит
        последнюю позицию каждого `icao24` вечно, и без неё на карте навсегда
        зависают борта, севшие часы назад.
      parameters:
        - $ref: "#/components/parameters/LonMin"
        - $ref: "#/components/parameters/LatMin"
        - $ref: "#/components/parameters/LonMax"
        - $ref: "#/components/parameters/LatMax"
      responses:
        "200":
          description: >-
            Массив позиций. Пустой массив — валидный ответ и **не** признак
            ошибки: так же выглядит и пустой участок карты, и остановленный
            пайплайн. Различить их можно только через `GET /pipeline-status`.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flight"
        "503":
          description: >-
            YTsaurus недоступен или вернул ошибку. Поле `message` заполнено
            по-русски и пригодно для показа пользователю.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /flights/{icao24}:
    get:
      tags: [flights]
      summary: Последняя известная позиция борта
      description: >-
        Карточка борта по клику на карте. Тело ответа — тот же объект, что
        и элемент массива в `GET /flights/live`.


        Трек сюда **не входит**, за ним нужен отдельный запрос
        к `GET /flights/{icao24}/track`.
      parameters:
        - $ref: "#/components/parameters/Icao24"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flight"
        "404":
          description: >-
            Борта нет в выдаче. Два неразличимых по ответу случая: `icao24`
            не прошёл проверку формата (не 6 hex-символов) либо борта нет
            в `positions_current` — он не наблюдался дольше окна свежести.
            Отдельного 400 на кривой формат нет.


            Поле `message` заполнено по-русски и пригодно для показа
            пользователю.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: >-
            YTsaurus недоступен или вернул ошибку. Поле `message` заполнено
            по-русски и пригодно для показа пользователю.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /flights/{icao24}/track:
    get:
      tags: [flights]
      summary: Трек борта из накопленной истории
      description: >-
        Точки строятся по `positions_history` и приходят **уже
        в хронологическом порядке** — сортировать на клиенте не нужно.


        Плотность точек — одна на опрос пайплайна, то есть примерно раз
        в 15 секунд (около 240 точек за час). В паузах пайплайна в треке будут
        разрывы, поэтому соединять прямой линией далеко разнесённые по времени
        точки не стоит.
      parameters:
        - $ref: "#/components/parameters/Icao24"
        - name: sinceSeconds
          in: query
          required: false
          description: Глубина окна назад от текущего момента, в секундах
          schema:
            type: integer
            format: int64
            minimum: 1
            default: 3600
      responses:
        "200":
          description: >-
            Массив точек трека. В отличие от карточки борта, неизвестный или
            некорректный `icao24` даёт **пустой массив и 200**, а не 404.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackPoint"
        "503":
          description: >-
            YTsaurus недоступен или вернул ошибку. Поле `message` заполнено
            по-русски и пригодно для показа пользователю.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /airports:
    get:
      tags: [airports]
      summary: Список аэропортов из справочника OurAirports
      description: >-
        Справочник `ref_airports` (OurAirports) с фильтрацией, сортировкой
        и пагинацией — FR12. Загруженности, числа рейсов и индекса трафика
        здесь **нет**: считать их не из чего, в пайплайне нет ни одной таблицы
        рейсов.


        **Отдаются не все строки справочника.** В OurAirports 86 тысяч записей,
        из которых 23 тысячи вертодромов, 13 тысяч закрытых площадок, полторы
        тысячи гидроаэродромов и воздухоплавательных площадок. Наружу уходят
        только `large_airport`, `medium_airport` и `small_airport` — около
        48 тысяч. Отключить этот фильтр параметром нельзя: вертодром на карте
        рейсов не принимает, а список аэропортов он засоряет. Из-за него же
        `total` меньше `row_count` таблицы.


        **Первый запрос после старта сервиса занимает десятки секунд.**
        `ref_airports` — статическая таблица, `select_rows` по ней не работает
        («Table ... is not dynamic»), поэтому она читается целиком через
        `read_table`: ~20 МБ и порядка 40 секунд. Снапшот кладётся в память
        и живёт `YT_REF_AIRPORTS_CACHE_TTL_SECONDS` (сутки по умолчанию), так
        что последующие ответы — миллисекунды. Обратная сторона: перезалив
        справочника сервис заметит только после истечения TTL или рестарта.
      parameters:
        - name: search
          in: query
          required: false
          description: >-
            Подстрока без учёта регистра. Ищется по тем же полям, которые видны
            в ответе: `name`, `city`, `icao`, `iata`. Совпадение — именно
            подстрока, а не префикс: `sheremet` найдёт «Sheremetyevo
            International Airport».


            Подстановочных символов нет. Значение сравнивается в памяти
            и в запрос к YTsaurus не попадает, поэтому кавычка и процент —
            обычные символы: они просто ничему не соответствуют и дают пустой
            список, а не ошибку. Длиннее 100 символов значение обрезается.


            Поиск по коду аэропорта работает не для всех строк: `icao` — это
            `icao_code`, а если он пуст (88% записей OurAirports), то `ident`.
            У 785 записей заполнены оба и они различаются — такую строку
            найдёт `icao_code`, но не `ident`.
          schema:
            type: string
            example: "Sheremetyevo"
        - name: country
          in: query
          required: false
          description: >-
            Точное совпадение с `iso_country` — двухбуквенный код ISO 3166-1
            alpha-2, регистр не важен. Не подстрока: `RU` не найдёт `RUS`.
            Неизвестный код — пустой список и `total: 0`, а не 400.
          schema:
            type: string
            example: "RU"
        - name: sortBy
          in: query
          required: false
          description: >-
            Единственное поддержанное значение — `name` (по названию, без учёта
            регистра, ничьи разрешаются по `icao`). Значения `trafficIndex`
            и `totalFlights` из `docs/openapi.yaml` не реализованы: этих метрик
            в ответе нет.


            Без параметра порядок — как в таблице, то есть по `ident`
            по возрастанию; он детерминирован, поэтому пагинация без сортировки
            тоже не перемешивает страницы. Незнакомое значение молча
            игнорируется и даёт тот же табличный порядок.
          schema:
            type: string
            enum: [name]
        - name: page
          in: query
          required: false
          description: >-
            Номер страницы, начиная с 1. Значения меньше 1 подтягиваются к 1,
            а не отвергаются. Страница за концом выборки — пустой `items`
            при неизменном `total` и с тем самым `page`, который запросили.
          schema:
            type: integer
            default: 1
            example: 1
        - name: pageSize
          in: query
          required: false
          description: >-
            Размер страницы. Значение зажимается в диапазон 1…500, то есть
            `pageSize=100000` вернёт 500 записей и `pageSize: 500` в ответе —
            запросить весь справочник одним ответом нельзя.
          schema:
            type: integer
            default: 50
            maximum: 500
            example: 50
        - $ref: "#/components/parameters/LonMin"
        - $ref: "#/components/parameters/LatMin"
        - $ref: "#/components/parameters/LonMax"
        - $ref: "#/components/parameters/LatMax"
        - name: limit
          in: query
          required: false
          description: >-
            Картовый режим: сколько аэропортов вернуть одним ответом.
            Заданный `limit` **отменяет** табличную пагинацию — `page`
            и `pageSize` в запросе игнорируются, а в ответе приходят
            `page: 1` и `pageSize`, равный фактическому лимиту.
            `total` при этом остаётся полным размером выборки после
            фильтров, так что по нему видно, сколько объектов не поместилось.
            Значение зажимается в диапазон 1…10000.
          schema:
            type: integer
            maximum: 10000
            example: 500
      responses:
        "200":
          description: >-
            Страница справочника. Пустой `items` — валидный ответ: так выглядят
            и слишком узкий фильтр, и страница за концом выборки.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AirportsListResponse"
        "400":
          description: >-
            Параметр не разобрался как число (`page=abc`, `latMin=север`).
            Осмысленные, но абсурдные значения (`page=0`, `pageSize=100000`,
            перевёрнутый bbox) под этот случай **не** подпадают — они
            обрабатываются молча. В `message` — техническое пояснение
            от Spring, по-английски.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: >-
            YTsaurus недоступен, вернул ошибку или таблицы `ref_airports` нет
            по настроенному пути. Пока снапшот жив в памяти, недоступность
            YTsaurus на эту ручку не влияет.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /pipeline-status:
    get:
      tags: [pipeline]
      summary: Состояние поставщика позиций
      description: >-
        Отвечает на вопрос «почему на карте пусто». Без этого эндпоинта
        отличить остановленный пайплайн от честно пустого участка карты
        невозможно — оба случая выглядят как `200 []`.


        Отвечает «едут ли данные вообще», но не «почему встали»: источник —
        таблица watermark'ов `pipeline_job_state`, причин остановки в ней нет.


        Достаточно поллить раз в 30 секунд, отдельно от позиций.
      responses:
        "200":
          description: >-
            Текущее состояние. Отвечает всегда, даже если пайплайн ни разу
            не запускался — тогда `status: unknown` и `stale: true`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineStatus"
        "503":
          description: >-
            YTsaurus недоступен или вернул ошибку. Поле `message` заполнено
            по-русски и пригодно для показа пользователю.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

components:
  parameters:
    Icao24:
      name: icao24
      in: path
      required: true
      description: >-
        ICAO24-адрес борта: ровно 6 hex-символов. Регистр не важен, сервер
        приводит значение к нижнему. Всё, что не подходит под формат,
        трактуется как «борт не найден».
      schema:
        type: string
        pattern: "^[0-9a-fA-F]{6}$"
        example: "01023b"
    LonMin:
      name: lonMin
      in: query
      required: false
      description: >-
        Западная граница видимой области карты.


        Параметры `lonMin`/`latMin`/`lonMax`/`latMax` задаются **только все
        четыре вместе**. Если задан не весь набор, фильтр молча игнорируется
        и вернётся весь мир — ошибки при этом не будет, но ответ окажется
        на порядок тяжелее ожидаемого.


        Проверок на осмысленность нет: перевёрнутая рамка (`lonMin > lonMax`
        или `latMin > latMax`) вернёт пустую выдачу, а не 400. Переход через
        антимеридиан (±180°) не поддержан — такой bbox тоже даст пусто.


        Один и тот же набор имён используют все ручки с картой; порядок
        «долгота, потом широта» взят из продуктового контракта
        `docs/openapi.yaml`, из которого фронтенд генерит типы.
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
        example: -6.0
    LatMin:
      name: latMin
      in: query
      required: false
      description: Южная граница видимой области карты
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
        example: 41.0
    LonMax:
      name: lonMax
      in: query
      required: false
      description: Восточная граница видимой области карты
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
        example: 10.0
    LatMax:
      name: latMax
      in: query
      required: false
      description: Северная граница видимой области карты
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
        example: 52.0

  schemas:
    Flight:
      type: object
      description: >-
        Позиция борта из `positions_current`, обогащённая справочником ВС
        по `icao24`.


        Все поля всегда присутствуют в JSON; необязательность выражена через
        `nullable`, а не через отсутствие ключа.
      properties:
        icao24:
          type: string
          description: ICAO24-адрес, 6 hex-символов в нижнем регистре. Ключ борта
          example: "01023b"
        callsign:
          type: string
          nullable: true
          description: Позывной, уже обрезанный от пробелов
          example: "MSC2932"
        originCountry:
          type: string
          nullable: true
          description: Страна регистрации борта
          example: "Egypt"
        timePosition:
          type: integer
          format: int64
          description: Момент фиксации позиции, unix seconds
          example: 1786841273
        lat:
          type: number
          format: double
          description: Широта WGS-84, градусы
          example: 45.41
        lon:
          type: number
          format: double
          description: Долгота WGS-84, градусы
          example: 20.0
        baroAltitude:
          type: number
          format: double
          nullable: true
          description: Барометрическая высота в **метрах**
          example: 10957.5
        onGround:
          type: boolean
          description: Борт находится на земле
        velocity:
          type: number
          format: double
          nullable: true
          description: Путевая скорость в **м/с**
          example: 229.9
        trueTrack:
          type: number
          format: double
          nullable: true
          description: >-
            Курс в градусах от севера по часовой стрелке. Задаёт поворот
            иконки борта на карте
          example: 324.75
        manufacturername:
          type: string
          nullable: true
          description: >-
            Производитель из справочника ВС. Имя поля в одно слово и в нижнем
            регистре — так колонка называется в схеме YT
          example: "Airbus Industrie"
        model:
          type: string
          nullable: true
          description: Модель ВС из справочника
          example: "A320-251N"
        operator:
          type: string
          nullable: true
          description: >-
            Авиакомпания-эксплуатант. Борт, которого нет в справочнике ВС,
            всё равно возвращается (FR4) — тогда `manufacturername`, `model`
            и `operator` равны `null`
          example: "Air Cairo"
      required:
        - icao24
        - callsign
        - originCountry
        - timePosition
        - lat
        - lon
        - baroAltitude
        - onGround
        - velocity
        - trueTrack
        - manufacturername
        - model
        - operator

    TrackPoint:
      type: object
      description: Одна точка трека борта из `positions_history`
      properties:
        timePosition:
          type: integer
          format: int64
          description: Момент фиксации позиции, unix seconds
        lat:
          type: number
          format: double
          description: Широта WGS-84
        lon:
          type: number
          format: double
          description: Долгота WGS-84
        baroAltitude:
          type: number
          format: double
          nullable: true
          description: Барометрическая высота в метрах, позволяет красить трек по высоте
      required: [timePosition, lat, lon, baroAltitude]

    AirportsListResponse:
      type: object
      description: >-
        Страница справочника аэропортов. Конверт, в отличие от `/flights/live`,
        нужен: без `total` клиент не может нарисовать пагинацию.
      properties:
        asOf:
          type: integer
          format: int64
          description: >-
            Момент последней записи таблицы `ref_airports` (unix seconds) —
            `modification_time` узла в Кипарисе, а не время ответа сервиса.
            Справочник статический и сам по себе не пересчитывается, так что
            это единственная осмысленная отметка свежести: она показывает,
            когда пайплайн залил справочник, и не меняется от запроса
            к запросу. `0` — атрибут не прочитался.
          example: 1786793001
        items:
          type: array
          description: Аэропорты текущей страницы
          items:
            $ref: "#/components/schemas/Airport"
        page:
          type: integer
          description: >-
            Номер отданной страницы. Совпадает с запрошенным (после
            подтягивания значений < 1 к единице); в картовом режиме
            с `limit` всегда 1
          example: 1
        pageSize:
          type: integer
          description: >-
            Фактический размер страницы после зажатия в допустимый диапазон.
            В картовом режиме равен фактическому `limit`
          example: 50
        total:
          type: integer
          description: >-
            Сколько аэропортов подошло под фильтры — до пагинации и до `limit`,
            но уже **после** отсева нeаэропортовых типов
          example: 943
      required: [asOf, items, page, pageSize, total]

    Airport:
      type: object
      description: >-
        Запись справочника OurAirports. Полей `totalFlights24h`
        и `trafficIndex` из `docs/openapi.yaml` здесь нет: рейсов
        по аэропортам пайплайн не считает.
      properties:
        icao:
          type: string
          description: >-
            `icao_code` из OurAirports, а если он пуст — `ident`. Пустым
            не бывает, но и настоящим ICAO-кодом является не всегда: у 88%
            записей справочника `icao_code` не заполнен, и тогда сюда попадает
            локальный идентификатор вида `00AA` или `RU-0796`
          example: "UUEE"
        iata:
          type: string
          nullable: true
          description: IATA-код. `null` у 89% записей справочника — это норма, а не сбой
          example: "SVO"
        name:
          type: string
          description: Название аэропорта, как в OurAirports
          example: "Sheremetyevo International Airport"
        city:
          type: string
          nullable: true
          description: >-
            `municipality` из OurAirports. `null` примерно у 5% записей —
            населённый пункт в источнике не указан
          example: "Moscow"
        country:
          type: string
          nullable: true
          description: >-
            `iso_country`, код ISO 3166-1 alpha-2. По схеме таблицы поле
            необязательное, поэтому объявлено nullable, хотя в текущей
            выгрузке заполнено у всех строк
          example: "RU"
        position:
          $ref: "#/components/schemas/LatLon"
      required: [icao, iata, name, city, country, position]

    LatLon:
      type: object
      description: Точка WGS-84
      properties:
        lat:
          type: number
          format: double
          example: 55.976858
        lon:
          type: number
          format: double
          example: 37.41121
      required: [lat, lon]

    PipelineStatus:
      type: object
      description: >-
        Состояние пайплайна, собранное из таблицы `pipeline_job_state` —
        watermark'ов джобов вида `(job_name, watermark_ts, updated_at)`.
        Берётся джоб, отчитавшийся последним.


        Источник знает только, докуда доехали данные, поэтому единственный
        содержательный сигнал здесь — `stale`. Причину остановки он не
        различает: паузу по лимиту кредитов OpenSky, недоступность источника
        и упавший джоб внешне не отличить, `resumesAt` всегда `null`.
        Двухступенчатый UX («обновление приостановлено до HH:MM», затем
        «данные устарели») на этих данных не построить — для него нужен
        отдельный хартбит `ingest_service`.
      properties:
        status:
          type: string
          enum:
            - ok
            - unknown
          description: >-
            Отчитывался ли пайплайн вообще:


            * `ok` — в таблице есть строка джоба. Означает только это:
            «джоб зарегистрировал своё состояние». Работает ли он **сейчас**,
            показывает `stale`, а не этот код.


            * `unknown` — в `pipeline_job_state` нет ни одной строки: джобы
            ни разу не отчитывались либо таблицу пересоздали. Всегда приходит
            вместе со `stale: true`.
        lastSuccessAt:
          type: integer
          format: int64
          nullable: true
          description: >-
            Watermark последнего отчитавшегося джоба, unix seconds — граница,
            до которой данные обработаны. Считаем свежесть именно по нему,
            а не по времени записи: если джоб крутится, а источник умер,
            время записи растёт при стоящих данных. `null` — джоб
            зарегистрировался, но ещё ничего не обработал
        resumesAt:
          type: integer
          format: int64
          nullable: true
          description: >-
            Момент, когда опрос возобновится, unix seconds. С текущим
            источником **всегда `null`**: `pipeline_job_state` про паузы
            ничего не знает
        stale:
          type: boolean
          description: >-
            Данные считаются устаревшими: watermark не продвигался дольше
            `PIPELINE_STALE_AFTER_SECONDS` (по умолчанию 120 с).
            При `lastSuccessAt: null` всегда `true`
      required: [status, lastSuccessAt, resumesAt, stale]

    ApiError:
      type: object
      description: >-
        Единое тело ошибки сервиса: его формирует `ApiExceptionHandler` для
        всех кодов, включая 5xx. Поле `message` есть всегда — дефолтное тело
        Spring без него наружу больше не выходит.


        У доменных ошибок (404) и у отказа YTsaurus (503) `message` написан
        по-русски и годится для показа пользователю. У ошибок самого Spring MVC
        (400 на неразобранный параметр, 405) там техническое пояснение
        по-английски — его лучше заменить своим текстом на клиенте.
      properties:
        timestamp:
          type: string
          format: date-time
          description: Момент формирования ответа, ISO-8601
          example: "2026-08-18T09:14:22.481Z"
        status:
          type: integer
          description: HTTP-код, дублирует статус ответа
          example: 404
        error:
          type: string
          description: Стандартное название HTTP-статуса
          example: "Not Found"
        message:
          type: string
          description: Причина ошибки; пустым не бывает
          example: "Позиция для борта icao24=abc123 не найдена"
      required: [timestamp, status, error, message]