# backend/positions-service/src/main/java/com/skypulse/positions/api/dto/AirportsListResponse.java
package com.skypulse.positions.api.dto;

import java.util.List;

public record AirportsListResponse(
        long asOf,
        List<AirportDto> items,
        int page,
        int pageSize,
        int total
) {
}