# backend/positions-service/src/main/java/com/skypulse/positions/model/Airport.java
package com.skypulse.positions.model;


public record Airport(
        String icao,
        String iata,
        String name,
        String type,
        String city,
        String country,
        double lat,
        double lon
) {
}